pyproject.toml
[tool.poetry]
name = "commit-message-generation-openai"
version = "0.0.1"
description = ""
authors = ["Alexandra Eliseeva <alexandra.eliseeva@jetbrains.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.4"
tiktoken = "^0.3.3"
backoff = "^2.2.1"
hydra-core = "1.2.0"
torchmetrics = "0.9.3"
datasets = "2.4.0"
nltk = "3.6.4"
rouge-score = "0.0.4"
sacrebleu = "2.0.0"
rapidfuzz = "^2.15.1"
jsonlines = "^3.1.0"
wandb = "^0.14.2"

[tool.poetry.group.dev.dependencies]
black = "^22.6.0"
isort = "^5.10.1"
mypy = "^0.981"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
